extract model-selection helpers into providers/model-selection.ts
